MentorMate-ai/mentor-chatbot-genai/requirements.txt
requests
google-generativeai
pillow
python-dotenv

# pip install -r requirements.txt